+++ context/app/routes_main.py
@@ -156,3 +156,16 @@ def iframe_page(path):
         title=f'{path} iframe',
         flask_data=flask_data
     )
+
+
+@blueprint.route('/genes/<geneSymbol>')
+def genes(geneSymbol):
+    flask_data = {
+        **get_default_flask_data(),
+        'geneSymbol': geneSymbol
+    }
+    return render_template(
+        'base-pages/react-content.html',
+        title='genes',
+        flask_data=flask_data
+    )