+++ context/app/static/js/components/cells/DatasetsSelectedByExpression/hooks.js
@@ -1,7 +1,6 @@
-import React, { useState, useContext } from 'react';
-
+import React, { useState } from 'react';
 import CellsService from 'js/components/cells/CellsService';
-import { AppContext } from 'js/components/Providers';
+import { useAppContext } from 'js/components/Contexts';
 import { fetchSearchData } from 'js/hooks/useSearchData';
 import useCellsChartLoadingStore from 'js/stores/useCellsChartLoadingStore';
 import { useStore } from 'js/components/cells/store';
@@ -43,7 +42,7 @@ function getSearchQuery(cellsResults) {
       'uuid',
       'hubmap_id',
       'mapped_data_types',
-      'origin_sample.mapped_organ',
+      'origin_samples_unique_mapped_organs',
       'donor',
       'last_modified_timestamp',
     ],
@@ -60,7 +59,7 @@ function buildHitsMap(hits) {
 
 function useDatasetsSelectedByExpression({ completeStep }) {
   const [message, setMessage] = useState(null);
-  const { elasticsearchEndpoint, groupsToken } = useContext(AppContext);
+  const { elasticsearchEndpoint, groupsToken } = useAppContext();
   const [genomicModality, setGenomicModality] = useState('rna');
   const resetFetchedUUIDs = useCellsChartLoadingStore(chartsStoreSelector);