+++ context/app/static/js/components/organ/Samples/Samples.jsx
@@ -51,7 +51,7 @@ function Samples({ organTerms }) {
             {
               bool: {
                 should: organTerms.map((searchTerm) => ({
-                  term: { 'origin_sample.mapped_organ.keyword': searchTerm },
+                  term: { 'origin_samples.mapped_organ.keyword': searchTerm },
                 })),
               },
             },