+++ context/app/static/js/components/organ/utils.js
@@ -2,7 +2,7 @@ function getSearchURL({ entityType, organTerms, assay }) {
   const query = new URLSearchParams();
   query.set('entity_type[0]', entityType);
   organTerms.forEach((term, i) => {
-    query.set(`origin_sample.mapped_organ[${i}]`, term);
+    query.set(`origin_samples.mapped_organ[${i}]`, term);
   });
   if (assay) {
     query.set('mapped_data_types[0]', assay);