+++ context/app/static/js/pages/Dataset/Dataset.jsx
@@ -1,7 +1,7 @@
-import React, { useContext, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import SvgIcon from '@material-ui/core/SvgIcon';
 
-import { AppContext } from 'js/components/Providers';
+import { useAppContext } from 'js/components/Contexts';
 import { LightBlueLink } from 'js/shared-styles/Links';
 import Files from 'js/components/detailPage/files/Files';
 import ProvSection from 'js/components/detailPage/provenance/ProvSection';
@@ -52,7 +52,7 @@ function SummaryDataChildren({
   hubmap_id,
   uuid,
 }) {
-  const { isWorkspacesUser } = useContext(AppContext);
+  const { isWorkspacesUser } = useAppContext();
 
   const createDatasetWorkspace = useDatasetWorkspace({ entity_type, uuid });
 
@@ -95,12 +95,11 @@ function DatasetDetail({ assayMetadata, vitData, hasNotebook, visLiftedUUID }) {
     metadata,
     files,
     donor,
-    source_sample,
+    source_samples,
     uuid,
     data_types,
     mapped_data_types,
-    origin_sample,
-    origin_sample: { mapped_organ },
+    origin_samples,
     group_name,
     created_by_user_displayname,
     created_by_user_email,
@@ -120,9 +119,13 @@ function DatasetDetail({ assayMetadata, vitData, hasNotebook, visLiftedUUID }) {
   } = assayMetadata;
   const isLatest = !('next_revision_uuid' in assayMetadata);
 
+  // TODO: Update design to reflect samples and datasets which have multiple origin samples with different organs.
+  const origin_sample = origin_samples[0];
+  const { mapped_organ } = origin_sample;
+
   const combinedStatus = getCombinedDatasetStatus({ sub_status, status });
 
-  const combinedMetadata = combineMetadata(donor, origin_sample, source_sample, metadata);
+  const combinedMetadata = combineMetadata(donor, origin_sample, source_samples, metadata);
 
   const collectionsData = useDatasetsCollections([uuid]);